refactor course data search to remove duplicate code
