+++ api/controllers/search-controller.js
@@ -1,10 +1,16 @@
 'use-strict';
 
 var searchService = require("../services/search-service");
+var endpoints = searchService.ENDPOINTS;
 
-exports.searchCourse = function (req, res) {
-    searchService.sendQueryToGriddy(req.query.searchQuery, function (data) {
-        var matchingCourses = searchService.parseGriddyResponse(data);
-        res.send(matchingCourses);
+exports.courseSearch = function (req, res) {
+    searchService.sendQueryToGriddy(endpoints.courseSearch, req.query.searchQuery, function (data) {
+        res.send(searchService.parseCourseSearchResults(data));
+    });
+};
+
+exports.courseDataSearch = function (req, res) {
+    searchService.sendQueryToGriddy(endpoints.courseDataSearch, req.query.courseId, function (data) {
+        res.send(searchService.parseCourseData(data));
     });
 };
