+++ api/routes/optiplan-routes.js
@@ -2,17 +2,15 @@
 
 module.exports = function (app) {
     var timetableController = require('../controllers/timetable-controller'),
-        searchController = require('../controllers/search-controller'),
-        courseDataController = require('../controllers/course-data-controller');
-
+        searchController = require('../controllers/search-controller');
     // set routes 
     app.route("/timetable")
         .get(timetableController.generateTimetable)
         .post(timetableController.saveTimetable);
 
     app.route("/search-course")
-        .get(searchController.searchCourse)
+        .get(searchController.courseSearch)
 
     app.route("/course")
-        .get(courseDataController.getCourseData)
-}
+        .get(searchController.courseDataSearch)
+};
